Sharding for the lazylru cache (#15)
The `LazyLRU` works hard to reduce blocking by reducing reorder operations, but there are limits. Another common mechanism to reduce blocking is to shard the cache. Two keys that target separate shards will not conflict, at least not from a memory-safety perspective. If lock contention is a problem you have in your application, sharding may be for you! There are volumes of literature on this subject and this PR means you _can_ shard your cache, but doesn't say whether or not you should.

The implementation here is a wrapper over the `lazylru.LazyLRU` cache. There are no attempts to spread out reaper threads or assist with memory locality. As a result, high shard counts will hurt performance. Considering that lock contention should reduce directly by the shard count, there are no advantages to high shard counts.