Making lazylru generic, deprecating the lazylru/generic package (#25)
This was supposed to happen last September. There is no need to support Go <=1.17 at this point, so the generic interface is sufficient for everyone. This will cause deprecation warnings for everyone using the `github.com/TriggerMail/lazylru/genric`, but switching to the `github.com/TriggerMail/lazylru` package should be seamless.